requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
torch>=2.0.0
torchvision>=0.15.0

# Natural Language Processing
nltk>=3.8.0
spacy>=3.6.0
transformers>=4.30.0

# Feature Engineering
category-encoders>=2.6.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Interpretation
shap>=0.42.0
lime>=0.2.0

# Deployment
streamlit>=1.25.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
joblib>=1.3.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# Data Validation
pydantic>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
